requirements.txt
matplotlib
mplfinance
numpy
pandas
ta
tqdm
python-dotenv
flask